Refactor Ray Serve behind a backend
